Adding gitignore and start of new dungeon generation system
Trying a different technique for a cleaner project. First go is using some Perlin Noise to generate a level with a bit of added clean-up. However, results are too unpredictable and can easily add lots of complexity for different clean-up rules so there aren't chunks sitting in space or other unreachable areas.